+++ src/JavaDependenciesManager.ts
@@ -17,26 +17,25 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import * as kamelCli from './kamel';
 import * as utils from './CamelKJSONUtils';
+import { execSync } from 'child_process';
 
 const PREFERENCE_KEY_JAVA_REFERENCED_LIBRARIES = "java.project.referencedLibraries";
 export const CAMEL_VERSION = "3.20.1";
 
 export async function initializeJavaDependenciesManager(context: vscode.ExtensionContext): Promise<void> {
-	const destination = parentDestinationFolderForDependencies(context);
+	const destination = destinationFolderForDependencies(context);
 	await downloadCommonCamelKJavaDependencies(context);
-	await initializeJavaSettingManagement(path.join(destination, 'dependencies'));
+	await initializeJavaSettingManagement(destination);
 }
 
 async function downloadJavaDependencies(context: vscode.ExtensionContext, groupId: string, artifactId: string, version: string) {
-	const destination = parentDestinationFolderForDependencies(context);
+	const destination = destinationFolderForDependencies(context);
 	const pomTemplate = context.asAbsolutePath(path.join('resources', 'maven-project', 'pom-to-copy-java-dependencies.xml'));
-	const directDependenciesDestination = path.join(destination, 'dependencies');
-	fs.mkdirSync(directDependenciesDestination, { recursive: true });
+	fs.mkdirSync(destination, { recursive: true });
 
 	const mvn = require('maven').create({
-		cwd: directDependenciesDestination,
+		cwd: destination,
 		file: pomTemplate
 	});
 
@@ -45,9 +44,9 @@ async function downloadJavaDependencies(context: vscode.ExtensionContext, groupI
 			'groupId': groupId,
 			'artifactId': artifactId,
 			'version': version,
-			'outputDirectory': directDependenciesDestination
+			'outputDirectory': destination
 		});
-	return directDependenciesDestination;
+	return destination;
 }
 
 async function initializeJavaSettingManagement(destination: string) {
@@ -74,20 +73,13 @@ export async function downloadSpecificCamelKJavaDependencies(
 		uri = vscode.window.activeTextEditor?.document.uri;
 	}
 	if (uri) {
-		const kamelLocal = kamelCli.create();
-		const parentDestination = parentDestinationFolderForDependencies(context);
-		await clearDestinationFolder(mainOutputChannel, parentDestination);
-		const command = `local build --integration-directory "${parentDestination}" "${uri.path}"`;
+		const destination = destinationFolderForDependencies(context);
+		await clearDestinationFolder(mainOutputChannel, destination);
+		const command = `jbang camel@apache/camel dependency copy --output-directory="${destination}" "${uri.path}"`;
 		try {
-			await kamelLocal.invoke(command);
+			execSync(command);
 			triggerRefreshOfJavaClasspath(context);
 		} catch(error) {
-			if(error instanceof Error) {
-				if(error.message.includes('unknown flag: --maven-repository')) {
-					utils.shareMessage(mainOutputChannel, 'A newer version of kamel CLI must be used to refresh classpath. 1.4+ is required. Either use a newer version and called again the refresh classpath action or restart VS Code to get back to basic dependencies available in classpath.');
-					return;
-				}
-			}
 			utils.shareMessage(mainOutputChannel, `Error while trying to refresh Java classpath based on file ${uri.path}:\n${error}`);
 		}
 	} else {
@@ -112,11 +104,6 @@ function triggerRefreshOfJavaClasspath(context: vscode.ExtensionContext) {
 	**/
 	updateReferencedLibraries("", destination);
 	updateReferenceLibraries(vscode.window.activeTextEditor, destination);
-}
-
-export function parentDestinationFolderForDependencies(context: vscode.ExtensionContext) {
-	const extensionStorage = context.globalStoragePath;
-	return  path.join(extensionStorage, `java-dependencies-${CAMEL_VERSION}`);
 }
 
 export function updateReferenceLibraries(editor: vscode.TextEditor | undefined, destination:string) {
@@ -164,5 +151,6 @@ function ensureReferencedLibrariesContainsCamelK(refLibrariesConfig: string[], c
     }
 }
 export function destinationFolderForDependencies(context: vscode.ExtensionContext) {
-	return path.join(parentDestinationFolderForDependencies(context), 'dependencies');
+	const extensionStorage = context.globalStoragePath;
+	return  path.join(extensionStorage, `java-dependencies`);
 }